Trait merino_suggest_traits::domain::CacheInputs

pub trait CacheInputs {
    fn add(&mut self, input: &[u8]);
    fn hash(&self) -> String;
}

Gathers inputs to be hashed to determine a cache key.

Required Methods

fn add(&mut self, input: &[u8])

Add data to the cache key.

fn hash(&self) -> String

Generate a cache key from the collected inputs so far.