Struct merino_suggest_traits::Proportion

pub struct Proportion(u32);

Represents a value from 0.0 to 1.0, inclusive. That is: a portion of something that cannot be negative or exceed 100%.

Stored internally as a u32.

Tuple Fields
0: u32
Implementations

impl Proportion

pub fn zero() -> Self

The lowest value for a portion, corresponding to 0%.

pub fn one() -> Self

The highest value for a portion, corresponding to 100%.

pub fn from<T>(v: T) -> Self where
T: TryInto<Self>,
<T as TryInto<Self>>::Error: Debug,

Converts a float value into a Proportion. Panics if the value is not between zero and one.

This is not implemented using [std::config::From] because you cannot implement both Try and TryFrom for the same pair of types, due to a blanket impl TryFor<T> for U where U: Try<T>.