Crate merino_suggest_traits

Suggestion backends for Merino.

Modules

device_info

Data structures that model information about a requester’s device. Device form factor, operating system, and browser are captured.

domain 🔒

Datatypes to better represent the domain of Merino.

metrics

Additional tools for recording metrics from suggestion providers.

Structs

NullProvider

A basic provider that never returns any suggestions.

Proportion

Represents a value from 0.0 to 1.0, inclusive. That is: a portion of something that cannot be negative or exceed 100%.

Suggestion

A suggestion to provide to a user.

SuggestionRequest

A request for suggestions.

SuggestionResponse

A response of suggestions, along with related metadata.

SupportedLanguages

Languages supported by the client.

Enums

CacheStatus

The relation between an object and a cache.

SetupError

Errors that may occur while setting up the provider.

SuggestError

Errors that may occur while querying for suggestions.

Constants

FIREFOX_TEST_VERSIONS

A provider that never provides any suggestions. The range of major Firefox version numbers to use for testing.

Traits

CacheInputs

Gathers inputs to be hashed to determine a cache key.

SuggestionProvider

A backend that can provide suggestions for queries.

Functions

convert_config

Convert a serde_json::Value into an appropriate provider configuration object.

fake_example_url 🔒

Helper to generate a URL to use for testing, of the form “https://example.com/fake#some-random-words”.

reconfigure_or_remake

Reconfigure the passed provider, or if there is a problem reconfiguring it, replace it with a newly created provider.

Type Definitions

MakeFreshType

A type that represents an object-safe function that can be passed to suggestion reconfigure methods to make new inner providers.